+++ sql/emaj--4.5.0--devel.sql
@@ -3122,6 +3122,602 @@ $_rlbk_init$
   END;
 $_rlbk_init$;
 
+CREATE OR REPLACE FUNCTION emaj._rlbk_planning(p_rlbkId INT)
+RETURNS INT LANGUAGE plpgsql
+SECURITY DEFINER SET search_path = pg_catalog, pg_temp AS
+$_rlbk_planning$
+-- This function builds the rollback steps for a rollback operation.
+-- It stores the result into the emaj_rlbk_plan table.
+-- The function returns the effective number of tables to process.
+-- It is called to perform a rollback operation. It is also called to simulate a rollback operation and get its duration estimate.
+-- It is called in an autonomous dblink transaction, if possible.
+-- The function is defined as SECURITY DEFINER so that emaj_viewer role can write into rollback tables, when estimating the rollback
+--   duration, without having specific privileges on them to do it.
+  DECLARE
+    v_groupNames             TEXT[];
+    v_mark                   TEXT;
+    v_isLoggedRlbk           BOOLEAN;
+    v_nbSession              INT;
+    v_nbSequence             INT;
+    v_ctrlStepName           emaj._rlbk_step_enum;
+    v_markTimeId             BIGINT;
+    v_avg_row_rlbk           INTERVAL;
+    v_avg_row_del_log        INTERVAL;
+    v_avg_fkey_check         INTERVAL;
+    v_fixed_step_rlbk        INTERVAL;
+    v_fixed_dblink_rlbk      INTERVAL;
+    v_fixed_table_rlbk       INTERVAL;
+    v_effNbTable             INT;
+    v_isEmajExtension        BOOLEAN;
+    v_batchNumber            INT;
+    v_checks                 INT;
+    v_estimDuration          INTERVAL;
+    v_estimDurationRlbkSeq   INTERVAL;
+    v_estimMethod            INT;
+    v_estimDropFkDuration    INTERVAL;
+    v_estimDropFkMethod      INT;
+    v_estimSetFkDefDuration  INTERVAL;
+    v_estimSetFkDefMethod    INT;
+    v_sessionLoad            INTERVAL[];
+    v_minSession             INT;
+    v_minDuration            INTERVAL;
+    v_nbStep                 INT;
+    v_fkList                 TEXT;
+    r_tbl                    RECORD;
+    r_fk                     RECORD;
+    r_batch                  RECORD;
+  BEGIN
+-- Get the rollback characteristics for the emaj_rlbk event.
+    SELECT rlbk_groups, rlbk_mark, rlbk_is_logged, rlbk_nb_session, rlbk_nb_sequence,
+           CASE WHEN rlbk_is_dblink_used THEN 'CTRL+DBLINK'::emaj._rlbk_step_enum ELSE 'CTRL-DBLINK'::emaj._rlbk_step_enum END
+      INTO v_groupNames, v_mark, v_isLoggedRlbk, v_nbSession, v_nbSequence,
+           v_ctrlStepName
+      FROM emaj.emaj_rlbk
+      WHERE rlbk_id = p_rlbkId;
+-- Get some mark attributes from emaj_mark.
+    SELECT mark_time_id INTO v_markTimeId
+      FROM emaj.emaj_mark
+      WHERE mark_group = v_groupNames[1]
+        AND mark_name = v_mark;
+-- Get all duration parameters that will be needed later from the emaj_param table, or get default values for rows
+-- that are not present in emaj_param table.
+    SELECT coalesce ((SELECT param_value_interval FROM emaj.emaj_param
+                        WHERE param_key = 'avg_row_rollback_duration'),'100 microsecond'::INTERVAL),
+           coalesce ((SELECT param_value_interval FROM emaj.emaj_param
+                        WHERE param_key = 'avg_row_delete_log_duration'),'10 microsecond'::INTERVAL),
+           coalesce ((SELECT param_value_interval FROM emaj.emaj_param
+                        WHERE param_key = 'avg_fkey_check_duration'),'5 microsecond'::INTERVAL),
+           coalesce ((SELECT param_value_interval FROM emaj.emaj_param
+                        WHERE param_key = 'fixed_step_rollback_duration'),'2.5 millisecond'::INTERVAL),
+           coalesce ((SELECT param_value_interval FROM emaj.emaj_param
+                        WHERE param_key = 'fixed_dblink_rollback_duration'),'4 millisecond'::INTERVAL),
+           coalesce ((SELECT param_value_interval FROM emaj.emaj_param
+                        WHERE param_key = 'fixed_table_rollback_duration'),'1 millisecond'::INTERVAL)
+           INTO v_avg_row_rlbk, v_avg_row_del_log, v_avg_fkey_check, v_fixed_step_rlbk, v_fixed_dblink_rlbk, v_fixed_table_rlbk;
+-- Process the sequences, if any in the tables groups.
+    IF v_nbSequence > 0 THEN
+-- Compute the cost for each RLBK_SEQUENCES step and keep it for later.
+      SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDurationRlbkSeq
+        FROM emaj._estimate_rlbk_step_duration('RLBK_SEQUENCES', NULL, NULL, NULL, v_nbSequence, v_fixed_step_rlbk, v_fixed_table_rlbk);
+-- Insert a RLBK_SEQUENCES step into emaj_rlbk_plan.
+-- Assign it the first session, so that it will be executed by the same session as the start mark set when the rollback is logged.
+      INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_session, rlbp_batch_number,
+                                       rlbp_estimated_quantity, rlbp_estimated_duration, rlbp_estimate_method)
+        VALUES (p_rlbkId, 'RLBK_SEQUENCES', '', '', '', 1, 1,
+                v_nbSequence, v_estimDurationRlbkSeq, v_estimMethod);
+    END IF;
+-- Insert into emaj_rlbk_plan a RLBK_TABLE step per table to effectively rollback.
+-- The numbers of log rows is computed using the _log_stat_tbl() function.
+-- A final check will be performed after tables will be locked to be sure no new table will have been updated.
+    INSERT INTO emaj.emaj_rlbk_plan
+            (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_is_repl_role_replica, rlbp_target_time_id,
+             rlbp_estimated_quantity)
+      SELECT p_rlbkId, 'RLBK_TABLE', rel_schema, rel_tblseq, '', FALSE, greatest(v_markTimeId, lower(rel_time_range)),
+             emaj._log_stat_tbl(t, greatest(v_markTimeId, lower(rel_time_range)), NULL)
+        FROM
+          (SELECT *
+             FROM emaj.emaj_relation
+             WHERE upper_inf(rel_time_range)
+               AND rel_group = ANY (v_groupNames)
+               AND rel_kind = 'r'
+          ) AS t
+        WHERE emaj._log_stat_tbl(t, greatest(v_markTimeId, lower(rel_time_range)), NULL) > 0;
+    GET DIAGNOSTICS v_effNbTable = ROW_COUNT;
+-- If nothing has to be rolled back, return quickly
+    IF v_nbSequence = 0 AND v_effNbTable = 0 THEN
+      RETURN 0;
+    END IF;
+-- Insert into emaj_rlbk_plan a LOCK_TABLE step per table currently belonging to the tables groups to process.
+    INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_is_repl_role_replica)
+      SELECT p_rlbkId, 'LOCK_TABLE', rel_schema, rel_tblseq, '', FALSE
+        FROM emaj.emaj_relation
+        WHERE upper_inf(rel_time_range)
+          AND rel_group = ANY(v_groupNames)
+          AND rel_kind = 'r';
+-- For tables to effectively rollback, add related steps (for FK, triggers, E-Maj logs) and adjust step properties.
+    IF v_effNbTable > 0 THEN
+-- Set the rlbp_is_repl_role_replica flag to TRUE for tables having all foreign keys linking tables:
+--   1) in the rolled back groups and 2) with the same rollback target mark.
+-- This only concerns emaj installed as an extension because one needs to be sure that the _rlbk_tbl() function is executed with a
+-- superuser role (this is needed to set the session_replication_role to 'replica').
+      v_isEmajExtension = EXISTS (SELECT 1 FROM pg_catalog.pg_extension WHERE extname = 'emaj');
+      IF v_isEmajExtension THEN
+        WITH fkeys AS (
+            -- the foreign keys belonging to tables to rollback
+            SELECT rlbp_schema, rlbp_table, c.conname, nf.nspname, tf.relname,
+                   (rel_group IS NOT NULL AND rel_group = ANY (v_groupNames) AND
+                    rlbp_target_time_id = greatest(v_markTimeId, lower(rel_time_range)))
+                     AS are_both_tables_in_groups_with_the_same_target_mark
+                     -- (rel_group IS NOT NULL AND rel_group = ANY (v_groupNames)) AS are_both_tables_in_groups,
+                     -- rlbp_target_time_id = greatest(v_markTimeId, lower(rel_time_range)) AS have_both_tables_the_same_target_mark
+              FROM emaj.emaj_rlbk_plan,
+                   pg_catalog.pg_constraint c
+                   JOIN pg_catalog.pg_class t ON (t.oid = c.conrelid)
+                   JOIN pg_catalog.pg_namespace n ON (n.oid = t.relnamespace)
+                   JOIN pg_catalog.pg_class tf ON (tf.oid = c.confrelid)
+                   JOIN pg_catalog.pg_namespace nf ON (nf.oid = tf.relnamespace)
+                   LEFT OUTER JOIN emaj.emaj_relation ON (rel_schema = nf.nspname AND rel_tblseq = tf.relname
+                                                          AND upper_inf(rel_time_range))
+              WHERE rlbp_rlbk_id = p_rlbkId                               -- The RLBK_TABLE steps for this rollback operation
+                AND rlbp_step = 'RLBK_TABLE'
+                AND contype = 'f'                                         -- FK constraints
+                AND tf.relkind = 'r'                                      -- only constraints referencing true tables, ie. excluding
+                                                                          --   partitionned tables
+                AND t.relname = rlbp_table
+                AND n.nspname = rlbp_schema
+          UNION ALL
+            -- the foreign keys referencing tables to rollback
+            SELECT rlbp_schema, rlbp_table, c.conname, n.nspname, t.relname,
+                   (rel_group IS NOT NULL AND rel_group = ANY (v_groupNames) AND
+                    rlbp_target_time_id = greatest(v_markTimeId, lower(rel_time_range)))
+                     AS are_both_tables_in_groups_with_the_same_target_mark
+              FROM emaj.emaj_rlbk_plan,
+                   pg_catalog.pg_constraint c
+                   JOIN pg_catalog.pg_class t ON (t.oid = c.conrelid)
+                   JOIN pg_catalog.pg_namespace n ON (n.oid = t.relnamespace)
+                   JOIN pg_catalog.pg_class tf ON (tf.oid = c.confrelid)
+                   JOIN pg_catalog.pg_namespace nf ON (nf.oid = tf.relnamespace)
+                   LEFT OUTER JOIN emaj.emaj_relation ON (rel_schema = n.nspname AND rel_tblseq = t.relname
+                                                          AND upper_inf(rel_time_range))
+              WHERE rlbp_rlbk_id = p_rlbkId                               -- The RLBK_TABLE steps for this rollback operation
+                AND rlbp_step = 'RLBK_TABLE'
+                AND contype = 'f'                                         -- FK constraints
+                AND t.relkind = 'r'                                       -- only constraints referenced by true tables, ie. excluding
+                                                                          --   partitionned tables
+                AND tf.relname = rlbp_table
+                AND nf.nspname = rlbp_schema
+        ), fkeys_agg AS (
+          -- aggregated foreign keys by tables to rollback
+          SELECT rlbp_schema, rlbp_table,
+                 count(*) AS nb_fk,
+                 count(*) FILTER (WHERE are_both_tables_in_groups_with_the_same_target_mark) AS nb_fk_ok
+            FROM fkeys
+            GROUP BY 1,2
+        )
+        UPDATE emaj.emaj_rlbk_plan
+          SET rlbp_is_repl_role_replica = TRUE
+          FROM fkeys_agg
+          WHERE rlbp_rlbk_id = p_rlbkId                                    -- The RLBK_TABLE steps for this rollback operation
+            AND rlbp_step IN ('RLBK_TABLE', 'LOCK_TABLE')
+            AND emaj_rlbk_plan.rlbp_table = fkeys_agg.rlbp_table
+            AND emaj_rlbk_plan.rlbp_schema = fkeys_agg.rlbp_schema
+            AND nb_fk = nb_fk_ok                                           -- all fkeys are linking tables 1) in the rolled back groups
+                                                                           -- and 2) with the same rollback target mark
+        ;
+      END IF;
+--
+-- Group tables into batchs to process all tables linked by foreign keys as a batch.
+--
+-- Start at 2, 1 being allocated to the RLBK_SEQUENCES step, if exists.
+      v_batchNumber = 2;
+-- Allocate tables with rows to rollback to batch number starting with the heaviest to rollback tables as reported by the
+-- emaj_log_stat_group() function.
+      FOR r_tbl IN
+        SELECT rlbp_schema, rlbp_table, rlbp_is_repl_role_replica
+          FROM emaj.emaj_rlbk_plan
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'RLBK_TABLE'
+          ORDER BY rlbp_estimated_quantity DESC, rlbp_schema, rlbp_table
+      LOOP
+-- If the table is not already allocated to a batch number (it may have been already allocated because of a fkey link).
+        IF EXISTS
+            (SELECT 0
+               FROM emaj.emaj_rlbk_plan
+               WHERE rlbp_rlbk_id = p_rlbkId
+                 AND rlbp_step = 'RLBK_TABLE'
+                 AND rlbp_schema = r_tbl.rlbp_schema
+                 AND rlbp_table = r_tbl.rlbp_table
+                 AND rlbp_batch_number IS NULL
+            ) THEN
+-- Allocate the table to the batch number, with all other tables linked by foreign key constraints.
+          PERFORM emaj._rlbk_set_batch_number(p_rlbkId, v_batchNumber, r_tbl.rlbp_schema, r_tbl.rlbp_table,
+                                              r_tbl.rlbp_is_repl_role_replica);
+          v_batchNumber = v_batchNumber + 1;
+        END IF;
+      END LOOP;
+--
+-- If unlogged rollback, register into emaj_rlbk_plan "disable log triggers", "deletes from log tables"
+-- and "enable log trigger" steps.
+--
+      IF NOT v_isLoggedRlbk THEN
+-- Compute the cost for each DIS_LOG_TRG step.
+        SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+          FROM emaj._estimate_rlbk_step_duration('DIS_LOG_TRG', NULL, NULL, NULL, NULL, v_fixed_step_rlbk, NULL);
+-- Insert all DIS_LOG_TRG steps.
+        INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_batch_number,
+                                         rlbp_estimated_duration, rlbp_estimate_method)
+          SELECT p_rlbkId, 'DIS_LOG_TRG', rlbp_schema, rlbp_table, '', rlbp_batch_number,
+                 v_estimDuration, v_estimMethod
+            FROM emaj.emaj_rlbk_plan
+            WHERE rlbp_rlbk_id = p_rlbkId
+              AND rlbp_step = 'RLBK_TABLE';
+-- Insert all DELETE_LOG steps. But the duration estimates will be computed later.
+-- The estimated number of log rows to delete is set to the estimated number of updates. This is underestimated in particular when
+-- SQL UPDATES are logged. But the collected statistics used for duration estimates are also based on the estimated number of updates.
+        INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_target_time_id,
+                                         rlbp_batch_number, rlbp_estimated_quantity)
+          SELECT p_rlbkId, 'DELETE_LOG', rlbp_schema, rlbp_table, '', rlbp_target_time_id, rlbp_batch_number, rlbp_estimated_quantity
+            FROM emaj.emaj_rlbk_plan
+            WHERE rlbp_rlbk_id = p_rlbkId
+              AND rlbp_step = 'RLBK_TABLE';
+-- Compute the cost for each ENA_LOG_TRG step.
+        SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+          FROM emaj._estimate_rlbk_step_duration('ENA_LOG_TRG', NULL, NULL, NULL, NULL, v_fixed_step_rlbk, NULL);
+-- Insert all ENA_LOG_TRG steps.
+        INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_batch_number,
+                                         rlbp_estimated_duration, rlbp_estimate_method)
+          SELECT p_rlbkId, 'ENA_LOG_TRG', rlbp_schema, rlbp_table, '', rlbp_batch_number, v_estimDuration, v_estimMethod
+            FROM emaj.emaj_rlbk_plan
+            WHERE rlbp_rlbk_id = p_rlbkId
+              AND rlbp_step = 'RLBK_TABLE';
+      END IF;
+--
+-- Process application triggers to temporarily set as ALWAYS triggers.
+-- This concerns triggers that must be kept enabled during the rollback processing but the rollback function for its table is executed
+-- with session_replication_role = replica.
+--
+-- Compute the cost for each SET_ALWAYS_APP_TRG step.
+      SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+        FROM emaj._estimate_rlbk_step_duration('SET_ALWAYS_APP_TRG', NULL, NULL, NULL, NULL, v_fixed_step_rlbk, NULL);
+-- Insert all SET_ALWAYS_APP_TRG steps.
+      INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_batch_number,
+                                       rlbp_estimated_duration, rlbp_estimate_method)
+        SELECT p_rlbkId, 'SET_ALWAYS_APP_TRG', rlbp_schema, rlbp_table, tgname, rlbp_batch_number, v_estimDuration, v_estimMethod
+          FROM emaj.emaj_rlbk_plan
+               JOIN pg_catalog.pg_class ON (relname = rlbp_table)
+               JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace AND nspname = rlbp_schema)
+               JOIN pg_catalog.pg_trigger ON (tgrelid = pg_class.oid)
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'RLBK_TABLE'                               -- rollback step
+            AND rlbp_is_repl_role_replica                              -- ... in session_replication_role = replica
+            AND NOT tgisinternal                                       -- application triggers only
+            AND tgname NOT IN ('emaj_trunc_trg','emaj_log_trg')
+            AND tgenabled = 'O'                                        -- ... enabled in local mode
+            AND EXISTS                                                 -- ... and to be kept enabled
+                  (SELECT 0
+                     FROM emaj.emaj_relation
+                     WHERE rel_schema = rlbp_schema
+                       AND rel_tblseq = rlbp_table
+                       AND upper_inf(rel_time_range)
+                       AND tgname = ANY (rel_ignored_triggers)
+                  );
+-- Compute the cost for each SET_LOCAL_APP_TRG step.
+      SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+        FROM emaj._estimate_rlbk_step_duration('SET_LOCAL_APP_TRG', NULL, NULL, NULL, NULL, v_fixed_step_rlbk, NULL);
+-- Insert all SET_LOCAL_APP_TRG steps
+      INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object,
+                                       rlbp_batch_number, rlbp_estimated_duration, rlbp_estimate_method)
+        SELECT p_rlbkId, 'SET_LOCAL_APP_TRG', rlbp_schema, rlbp_table, rlbp_object,
+               rlbp_batch_number, v_estimDuration, v_estimMethod
+          FROM emaj.emaj_rlbk_plan
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'SET_ALWAYS_APP_TRG';
+--
+-- Process application triggers to disable and re-enable.
+-- This concerns triggers that must be disabled during the rollback processing and the rollback function for its table is not executed
+-- with session_replication_role = replica.
+--
+-- Compute the cost for each DIS_APP_TRG step.
+      SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+        FROM emaj._estimate_rlbk_step_duration('DIS_APP_TRG', NULL, NULL, NULL, NULL, v_fixed_step_rlbk, NULL);
+-- Insert all DIS_APP_TRG steps.
+      INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_batch_number,
+                                       rlbp_estimated_duration, rlbp_estimate_method)
+        SELECT p_rlbkId, 'DIS_APP_TRG', rlbp_schema, rlbp_table, tgname, rlbp_batch_number, v_estimDuration, v_estimMethod
+          FROM emaj.emaj_rlbk_plan
+               JOIN pg_catalog.pg_class ON (relname = rlbp_table)
+               JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace AND nspname = rlbp_schema)
+               JOIN pg_catalog.pg_trigger ON (tgrelid = pg_class.oid)
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'RLBK_TABLE'                               -- rollback step
+            AND NOT tgisinternal                                       -- application triggers only
+            AND tgname NOT IN ('emaj_trunc_trg','emaj_log_trg')
+            AND (tgenabled IN ('A', 'R')                               -- enabled ALWAYS or REPLICA triggers
+                OR (tgenabled = 'O' AND NOT rlbp_is_repl_role_replica) -- or enabled ORIGIN triggers for rollbacks not processed
+                )                                                      --   in session_replication_role = replica)
+            AND NOT EXISTS                                             -- ... that must be disabled
+                  (SELECT 0
+                     FROM emaj.emaj_relation
+                     WHERE rel_schema = rlbp_schema
+                       AND rel_tblseq = rlbp_table
+                       AND upper_inf(rel_time_range)
+                       AND tgname = ANY (rel_ignored_triggers)
+                  );
+-- Compute the cost for each ENA_APP_TRG step.
+      SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+        FROM emaj._estimate_rlbk_step_duration('ENA_APP_TRG', NULL, NULL, NULL, NULL, v_fixed_step_rlbk, NULL);
+-- Insert all ENA_APP_TRG steps.
+      INSERT INTO emaj.emaj_rlbk_plan (rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object,
+                                       rlbp_app_trg_type,
+                                       rlbp_batch_number, rlbp_estimated_duration, rlbp_estimate_method)
+        SELECT p_rlbkId, 'ENA_APP_TRG', rlbp_schema, rlbp_table, rlbp_object,
+               CASE tgenabled WHEN 'A' THEN 'ALWAYS' WHEN 'R' THEN 'REPLICA' ELSE '' END,
+               rlbp_batch_number, v_estimDuration, v_estimMethod
+          FROM emaj.emaj_rlbk_plan
+               JOIN pg_catalog.pg_class ON (relname = rlbp_table)
+               JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace AND nspname = rlbp_schema)
+               JOIN pg_catalog.pg_trigger ON (tgrelid = pg_class.oid AND tgname = rlbp_object)
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'DIS_APP_TRG';
+--
+-- Process foreign key to define which action to perform on them
+--
+-- First compute the fixed duration estimates for each 'DROP_FK' and 'SET_FK_DEF' steps.
+      SELECT p_estimateMethod, p_estimatedDuration INTO v_estimDropFkMethod, v_estimDropFkDuration
+        FROM emaj._estimate_rlbk_step_duration('DROP_FK', NULL, NULL, NULL, NULL, v_fixed_step_rlbk, NULL);
+      SELECT p_estimateMethod, p_estimatedDuration INTO v_estimSetFkDefMethod, v_estimSetFkDefDuration
+        FROM emaj._estimate_rlbk_step_duration('SET_FK_DEF', NULL, NULL, NULL, NULL, v_fixed_step_rlbk, NULL);
+-- Select all foreign keys belonging to or referencing the tables to process.
+      FOR r_fk IN
+          SELECT c.oid AS conoid, c.conname, n.nspname, t.relname, t.reltuples, c.condeferrable,
+                 c.condeferred, c.confupdtype, c.confdeltype, r.rlbp_batch_number
+            FROM emaj.emaj_rlbk_plan r
+                 JOIN pg_catalog.pg_class t ON (t.relname = r.rlbp_table)
+                 JOIN pg_catalog.pg_namespace n ON (t.relnamespace  = n.oid AND n.nspname = r.rlbp_schema)
+                 JOIN pg_catalog.pg_constraint c ON (c.conrelid = t.oid)
+            WHERE c.contype = 'f'                                            -- FK constraints only
+              AND rlbp_rlbk_id = p_rlbkId
+              AND rlbp_step = 'RLBK_TABLE'                                   -- Tables to rollback
+              AND NOT rlbp_is_repl_role_replica                              -- ... not in a session_replication_role = replica
+        UNION
+          SELECT c.oid AS conoid, c.conname, n.nspname, t.relname, t.reltuples, c.condeferrable,
+                 c.condeferred, c.confupdtype, c.confdeltype, r.rlbp_batch_number
+            FROM emaj.emaj_rlbk_plan r
+                 JOIN pg_catalog.pg_class rt ON (rt.relname = r.rlbp_table)
+                 JOIN pg_catalog.pg_namespace rn ON (rn.oid = rt.relnamespace AND rn.nspname = r.rlbp_schema)
+                 JOIN pg_catalog.pg_constraint c ON (c.confrelid = rt.oid)
+                 JOIN pg_catalog.pg_class t ON (t.oid = c.conrelid)
+                 JOIN pg_catalog.pg_namespace n ON (n.oid = t.relnamespace)
+            WHERE c.contype = 'f'                                            -- FK constraints only
+              AND rlbp_rlbk_id = p_rlbkId
+              AND rlbp_step = 'RLBK_TABLE'                                   -- Tables to rollback
+              AND NOT rlbp_is_repl_role_replica                              -- ... not in a session_replication_role = replica
+            ORDER BY nspname, relname, conname
+      LOOP
+-- Depending on the foreign key characteristics, record as 'to be dropped' or 'to be set deferred' or 'to just be reset immediate'.
+        IF NOT r_fk.condeferrable OR r_fk.confupdtype <> 'a' OR r_fk.confdeltype <> 'a' THEN
+-- Non deferrable fkeys and fkeys with an action for UPDATE or DELETE other than 'no action' need to be dropped.
+          INSERT INTO emaj.emaj_rlbk_plan (
+            rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_batch_number,
+            rlbp_estimated_duration, rlbp_estimate_method
+            ) VALUES (
+            p_rlbkId, 'DROP_FK', r_fk.nspname, r_fk.relname, r_fk.conname, r_fk.rlbp_batch_number,
+            v_estimDropFkDuration, v_estimDropFkMethod
+            );
+          INSERT INTO emaj.emaj_rlbk_plan (
+            rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_batch_number, rlbp_object_def,
+            rlbp_estimated_quantity
+            ) VALUES (
+            p_rlbkId, 'ADD_FK', r_fk.nspname, r_fk.relname, r_fk.conname, r_fk.rlbp_batch_number, pg_get_constraintdef(r_fk.conoid),
+            r_fk.reltuples
+            );
+        ELSE
+-- Other deferrable but not deferred fkeys need to be set deferred.
+          IF NOT r_fk.condeferred THEN
+            INSERT INTO emaj.emaj_rlbk_plan (
+              rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_batch_number,
+              rlbp_estimated_duration, rlbp_estimate_method
+              ) VALUES (
+              p_rlbkId, 'SET_FK_DEF', r_fk.nspname, r_fk.relname, r_fk.conname, r_fk.rlbp_batch_number,
+              v_estimSetFkDefDuration, v_estimSetFkDefMethod
+              );
+          END IF;
+-- Deferrable fkeys are recorded as 'to be set immediate at the end of the rollback operation'.
+-- Compute the number of fkey values to check at set immediate time.
+          SELECT (coalesce(
+-- Get the number of rolled back rows in the referencing table, if any.
+             (SELECT rlbp_estimated_quantity
+                FROM emaj.emaj_rlbk_plan
+                WHERE rlbp_rlbk_id = p_rlbkId
+                  AND rlbp_step = 'RLBK_TABLE'                                   -- tables of the rollback event
+                  AND rlbp_schema = r_fk.nspname
+                  AND rlbp_table = r_fk.relname)                                 -- referencing schema.table
+              , 0)) + (coalesce(
+-- Get the number of rolled back rows in the referenced table, if any.
+             (SELECT rlbp_estimated_quantity
+                FROM emaj.emaj_rlbk_plan
+                     JOIN pg_catalog.pg_class rt ON (rt.relname = rlbp_table)
+                     JOIN pg_catalog.pg_namespace rn ON (rn.oid = rt.relnamespace AND rn.nspname = rlbp_schema)
+                     JOIN pg_catalog.pg_constraint c ON (c.confrelid  = rt.oid)
+                WHERE rlbp_rlbk_id = p_rlbkId
+                  AND rlbp_step = 'RLBK_TABLE'                                   -- tables of the rollback event
+                  AND c.oid = r_fk.conoid                                        -- constraint id
+             )
+              , 0)) INTO v_checks;
+-- And record the SET_FK_IMM step.
+          INSERT INTO emaj.emaj_rlbk_plan (
+            rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_batch_number, rlbp_estimated_quantity
+            ) VALUES (
+            p_rlbkId, 'SET_FK_IMM', r_fk.nspname, r_fk.relname, r_fk.conname, r_fk.rlbp_batch_number, v_checks
+            );
+        END IF;
+      END LOOP;
+-- Raise an exception if DROP_FK steps concerns inherited FK (i.e. FK set on a partitionned table)
+      SELECT string_agg(rlbp_schema || '.' || rlbp_table || '.' || rlbp_object, ', ')
+        INTO v_fkList
+        FROM emaj.emaj_rlbk_plan r
+             JOIN pg_catalog.pg_class t ON (t.relname = r.rlbp_table)
+             JOIN pg_catalog.pg_namespace n ON (t.relnamespace  = n.oid AND n.nspname = r.rlbp_schema)
+             JOIN pg_catalog.pg_constraint c ON (c.conrelid = t.oid AND c.conname = r.rlbp_object)
+        WHERE rlbp_rlbk_id = p_rlbkId
+          AND rlbp_step = 'DROP_FK'
+          AND coninhcount > 0;
+      IF v_fkList IS NOT NULL THEN
+        RAISE EXCEPTION '_rlbk_planning: Some foreign keys (%) would need to be temporarily dropped during the operation. '
+                        'But this would fail because they are inherited from a partitionned table.', v_fkList;
+      END IF;
+--
+-- Now compute the estimation duration for each complex step ('RLBK_TABLE', 'DELETE_LOG', 'ADD_FK', 'SET_FK_IMM').
+--
+-- Compute the rollback duration estimates for the tables.
+      FOR r_tbl IN
+        SELECT *
+          FROM emaj.emaj_rlbk_plan
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'RLBK_TABLE'
+      LOOP
+        SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+          FROM emaj._estimate_rlbk_step_duration('RLBK_TABLE', r_tbl.rlbp_schema, r_tbl.rlbp_table, NULL,
+                                                 r_tbl.rlbp_estimated_quantity, v_fixed_step_rlbk, v_avg_row_rlbk);
+        UPDATE emaj.emaj_rlbk_plan
+          SET rlbp_estimated_duration = v_estimDuration, rlbp_estimate_method = v_estimMethod
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'RLBK_TABLE'
+            AND rlbp_schema = r_tbl.rlbp_schema
+            AND rlbp_table = r_tbl.rlbp_table;
+      END LOOP;
+-- Compute the estimated log rows delete duration.
+      FOR r_tbl IN
+        SELECT *
+          FROM emaj.emaj_rlbk_plan
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'DELETE_LOG'
+      LOOP
+        SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+          FROM emaj._estimate_rlbk_step_duration('DELETE_LOG', r_tbl.rlbp_schema, r_tbl.rlbp_table, NULL,
+                                                 r_tbl.rlbp_estimated_quantity, v_fixed_step_rlbk, v_avg_row_del_log);
+        UPDATE emaj.emaj_rlbk_plan
+          SET rlbp_estimated_duration = v_estimDuration, rlbp_estimate_method = v_estimMethod
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'DELETE_LOG'
+            AND rlbp_schema = r_tbl.rlbp_schema
+            AND rlbp_table = r_tbl.rlbp_table;
+      END LOOP;
+-- Compute the fkey recreation duration.
+      FOR r_fk IN
+        SELECT *
+          FROM emaj.emaj_rlbk_plan
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'ADD_FK'
+      LOOP
+        SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+          FROM emaj._estimate_rlbk_step_duration('ADD_FK', r_tbl.rlbp_schema, r_tbl.rlbp_table, r_fk.rlbp_object,
+                                                 r_tbl.rlbp_estimated_quantity, v_fixed_step_rlbk, v_avg_fkey_check);
+        UPDATE emaj.emaj_rlbk_plan
+          SET rlbp_estimated_duration = v_estimDuration, rlbp_estimate_method = v_estimMethod
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'ADD_FK'
+            AND rlbp_schema = r_fk.rlbp_schema
+            AND rlbp_table = r_fk.rlbp_table
+            AND rlbp_object = r_fk.rlbp_object;
+      END LOOP;
+-- Compute the fkey checks duration.
+      FOR r_fk IN
+        SELECT * FROM emaj.emaj_rlbk_plan
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'SET_FK_IMM'
+      LOOP
+        SELECT p_estimateMethod, p_estimatedDuration INTO v_estimMethod, v_estimDuration
+          FROM emaj._estimate_rlbk_step_duration('SET_FK_IMM', r_tbl.rlbp_schema, r_tbl.rlbp_table, r_fk.rlbp_object,
+                                                 r_tbl.rlbp_estimated_quantity, v_fixed_step_rlbk, v_avg_fkey_check);
+        UPDATE emaj.emaj_rlbk_plan
+          SET rlbp_estimated_duration = v_estimDuration, rlbp_estimate_method = v_estimMethod
+          WHERE rlbp_rlbk_id = p_rlbkId
+            AND rlbp_step = 'SET_FK_IMM'
+            AND rlbp_schema = r_fk.rlbp_schema
+            AND rlbp_table = r_fk.rlbp_table
+            AND rlbp_object = r_fk.rlbp_object;
+      END LOOP;
+--
+-- Allocate batches to sessions to spread the load on sessions as best as possible.
+-- A batch represents all steps related to the processing of one table or several tables linked by foreign keys.
+--
+      IF v_nbSession = 1 THEN
+-- In single session rollback, assign all steps to session 1 at once.
+        UPDATE emaj.emaj_rlbk_plan
+          SET rlbp_session = 1
+          WHERE rlbp_rlbk_id = p_rlbkId;
+      ELSE
+-- Initialisation (for session 1, the RLBK_SEQUENCES step may have been already assigned).
+        v_sessionLoad [1] = coalesce(v_estimDurationRlbkSeq, '0 SECONDS'::INTERVAL);
+        FOR v_session IN 2 .. v_nbSession LOOP
+          v_sessionLoad [v_session] = '0 SECONDS'::INTERVAL;
+        END LOOP;
+-- Allocate tables batch to sessions, starting with the heaviest to rollback batch.
+        FOR r_batch IN
+          SELECT rlbp_batch_number, sum(rlbp_estimated_duration) AS batch_duration
+            FROM emaj.emaj_rlbk_plan
+            WHERE rlbp_rlbk_id = p_rlbkId
+              AND rlbp_batch_number IS NOT NULL
+              AND rlbp_session IS NULL
+            GROUP BY rlbp_batch_number
+            ORDER BY sum(rlbp_estimated_duration) DESC
+        LOOP
+-- Compute the least loaded session.
+          v_minSession = 1; v_minDuration = v_sessionLoad [1];
+          FOR v_session IN 2 .. v_nbSession LOOP
+            IF v_sessionLoad [v_session] < v_minDuration THEN
+              v_minSession = v_session;
+              v_minDuration = v_sessionLoad [v_session];
+            END IF;
+          END LOOP;
+-- Allocate the batch to the session.
+          UPDATE emaj.emaj_rlbk_plan
+            SET rlbp_session = v_minSession
+            WHERE rlbp_rlbk_id = p_rlbkId
+              AND rlbp_batch_number = r_batch.rlbp_batch_number;
+          v_sessionLoad [v_minSession] = v_sessionLoad [v_minSession] + r_batch.batch_duration;
+        END LOOP;
+      END IF;
+    END IF;
+-- Assign all not yet assigned 'LOCK_TABLE' steps to session 1.
+    UPDATE emaj.emaj_rlbk_plan
+      SET rlbp_session = 1
+      WHERE rlbp_rlbk_id = p_rlbkId
+        AND rlbp_session IS NULL;
+--
+-- Create the pseudo 'CTRL+DBLINK' or 'CTRL-DBLINK' step and compute its duration estimate.
+--
+-- Get the number of recorded steps (except LOCK_TABLE).
+    SELECT count(*) INTO v_nbStep
+      FROM emaj.emaj_rlbk_plan
+      WHERE rlbp_rlbk_id = p_rlbkId
+        AND rlbp_step <> 'LOCK_TABLE';
+    IF v_nbStep > 0 THEN
+-- If CTRLxDBLINK statistics are available, compute an average cost.
+      SELECT sum(rlbt_duration) * v_nbStep / sum(rlbt_quantity) INTO v_estimDuration
+        FROM emaj.emaj_rlbk_stat
+        WHERE rlbt_step = v_ctrlStepName
+          AND rlbt_quantity > 0;
+      v_estimMethod = 2;
+      IF v_estimDuration IS NULL THEN
+-- Otherwise, use the fixed_step_rollback_duration parameter.
+        v_estimDuration = v_fixed_dblink_rlbk * v_nbStep;
+        v_estimMethod = 3;
+      END IF;
+-- Insert the 'CTRLxDBLINK' pseudo step.
+      INSERT INTO emaj.emaj_rlbk_plan (
+          rlbp_rlbk_id, rlbp_step, rlbp_schema, rlbp_table, rlbp_object, rlbp_estimated_quantity,
+          rlbp_estimated_duration, rlbp_estimate_method
+        ) VALUES (
+          p_rlbkId, v_ctrlStepName, '', '', '', v_nbStep, v_estimDuration, v_estimMethod
+        );
+    END IF;
+-- Return the number of tables to effectively rollback.
+    RETURN v_effNbTable;
+  END;
+$_rlbk_planning$;
+
 CREATE OR REPLACE FUNCTION emaj._rlbk_session_lock(p_rlbkId INT, p_session INT)
 RETURNS VOID LANGUAGE plpgsql
 SECURITY DEFINER SET search_path = pg_catalog, pg_temp AS